// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  CUSTOMER: 'CUSTOMER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProductType = {
  READY_MADE: 'READY_MADE',
  CUSTOM_PRINT: 'CUSTOM_PRINT'
} as const

export type ProductType = (typeof ProductType)[keyof typeof ProductType]


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  PROCESSING: 'PROCESSING',
  PRINTING: 'PRINTING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const FileFormat = {
  STL: 'STL',
  THREE_MF: 'THREE_MF'
} as const

export type FileFormat = (typeof FileFormat)[keyof typeof FileFormat]


export const StockReason = {
  ORDER: 'ORDER',
  STOCK_PRINT: 'STOCK_PRINT',
  FAILURE: 'FAILURE',
  RESTOCK: 'RESTOCK',
  CORRECTION: 'CORRECTION'
} as const

export type StockReason = (typeof StockReason)[keyof typeof StockReason]